Clean up lapp-gen code and interface
Now the environment variable is LAPP_GEN, not LAPP_FILE, since it's more
than just a file, and when not passed a LAPP_GEN variable, it prints out
a usage text. Also, rather than writing a file itself with `struct`, it
simply prints to `stdout` so that users can redirect output where they
like.